requirements.txt
colorama==0.4.6
cryptography==43.0.1
python_bcrypt==0.3.2
